refactor: Consolidate test helpers into internal/testutil
- Move mockT and User structs, and related helper functions to internal/testutil package.
- Update all test files (assert_test.go, original_request_test.go, value_capture_comprehensive_test.go, example_test.go, phase2_demo_test.go, values_test.go) to import and use testutil.
- Update Go code snippets in README.md and docs/philosophy.md to reflect the changes.
- Ensure all tests pass after refactoring.
