+++ value_capture_demo_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/paveg/diagassert"
+	"github.com/paveg/diagassert/internal/testutil"
 )
 
 // ValueCaptureDemo demonstrates the enhanced output with value capture
@@ -15,17 +16,17 @@ func TestValueCaptureDemo(t *testing.T) {
 	t.Log("=== Demo of Value Capture API ===")
 
 	// Create a mock to capture output
-	mock := &mockT{}
+	mock := testutil.NewMockT()
 
 	// Demo 1: Simple value capture
 	t.Log("\n--- Demo 1: Simple Value Capture ---")
 	x := 15
 	diagassert.Assert(mock, x > 20, diagassert.V("x", x))
-	t.Logf("Output:\n%s", mock.getOutput())
+	t.Logf("Output:\n%s", mock.GetOutput())
 
 	// Demo 2: Multiple values with custom message
 	t.Log("\n--- Demo 2: Multiple Values with Custom Message ---")
-	mock = &mockT{}
+	mock = testutil.NewMockT()
 	age := 16
 	hasLicense := false
 	diagassert.Assert(mock, age >= 18 && hasLicense,
@@ -34,23 +35,23 @@ func TestValueCaptureDemo(t *testing.T) {
 			"hasLicense": hasLicense,
 		},
 		"User permission check failed")
-	t.Logf("Output:\n%s", mock.getOutput())
+	t.Logf("Output:\n%s", mock.GetOutput())
 
 	// Demo 3: Complex struct with method calls
 	t.Log("\n--- Demo 3: Complex Struct with Method Calls ---")
-	mock = &mockT{}
-	user := User{Name: "Alice", Age: 16}
+	mock = testutil.NewMockT()
+	user := testutil.User{Name: "Alice", Age: 16}
 	minAge := 18
 	diagassert.Assert(mock, user.Age >= minAge && user.IsAdult(),
 		diagassert.V("user", user),
 		diagassert.V("user.Age", user.Age),
 		diagassert.V("minAge", minAge),
 		"Age verification failed")
-	t.Logf("Output:\n%s", mock.getOutput())
+	t.Logf("Output:\n%s", mock.GetOutput())
 
 	// Demo 4: Mixed values and messages
 	t.Log("\n--- Demo 4: Mixed Values and Messages ---")
-	mock = &mockT{}
+	mock = testutil.NewMockT()
 	a, b, c := 10, 20, 30
 	diagassert.Assert(mock, a > b && b > c,
 		"First condition check",
@@ -59,5 +60,5 @@ func TestValueCaptureDemo(t *testing.T) {
 		diagassert.V("b", b),
 		diagassert.V("c", c),
 		"All values should form descending sequence")
-	t.Logf("Output:\n%s", mock.getOutput())
+	t.Logf("Output:\n%s", mock.GetOutput())
 }
